add login/registration form

+++ src/main/java/ru/privetdruk/wessenger/domain/Message.java
@@ -1,15 +1,12 @@
 package ru.privetdruk.wessenger.domain;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import javax.persistence.*;
 
 @Entity
 public class Message {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     public Message(String text, String tag) {
         this.text = text;
@@ -22,11 +19,11 @@ public class Message {
     private String text;
     private String tag;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
